+++ Lead-Server-Source/game/src/sectree_manager.cpp
@@ -643,7 +643,7 @@ const TMapRegion * SECTREE_MANAGER::GetMapRegion(long lMapIndex)
 	return NULL;
 }
 
-int SECTREE_MANAGER::GetMapIndex(long x, long y)
+int SECTREE_MANAGER::GetMapIndex(int32_t x, int32_t y)
 {
 	std::vector<TMapRegion>::iterator it = m_vec_mapRegion.begin();
 
@@ -790,7 +790,7 @@ bool SECTREE_MANAGER::GetMovablePosition(long lMapIndex, long x, long y, PIXEL_P
 	return false;
 }
 
-bool SECTREE_MANAGER::GetValidLocation(long lMapIndex, long x, long y, long & r_lValidMapIndex, PIXEL_POSITION & r_pos, BYTE empire)
+bool SECTREE_MANAGER::GetValidLocation(int32_t lMapIndex, int32_t x, int32_t y, int32_t & r_lValidMapIndex, PIXEL_POSITION & r_pos, BYTE empire)
 {
 	LPSECTREE_MAP pkSectreeMap = GetMap(lMapIndex);
